added collector_factory module for Quantum ESPRESSO data extraction
